Organize web component tests
